fix(web): move getHistoryItemHref off the client
Pure href helper was exported from a use-client module, so Next
turned it into a client reference. ProjectNeedsAttentionSection
called it during RSC render and threw (SOKOSUMI-RW).
